within time users can Vote

+++ app/Http/Controllers/Api/VoteController.php
@@ -46,44 +46,47 @@ class VoteController extends Controller
             if(auth()->id() <> $submitedChallenge->acceptedChallenge->user->id){ 
                 $data['message'] = 'The result of this Challenge is not based on Vote';
                 if($submitedChallenge->acceptedChallenge->challenge->result_type === 'vote'){
-                    $res = 200;
-                    $sub_id  = $submitedChallenge->id;
-                    $acceptedChallenges = $submitedChallenge->acceptedChallenge->challenge->acceptedChallenges;
-                    $isVoted = 0;
-                    $voters = Vote::where('user_id',auth()->id())->get();
-                    if($voters->first()) {
-                        foreach($acceptedChallenges as $acceptedChallenge){
-                            foreach($voters as $voter){
-                                if($voter->submited_challenge_id === $acceptedChallenge->submitChallenge->id){
-                                    $isVoted++;
+                    $data['message'] = 'You\'re out of time.';
+                    if(now() <= $submitedChallenge->acceptedChallenge->challenge->after_date){
+                        $res = 200;
+                        $sub_id  = $submitedChallenge->id;
+                        $acceptedChallenges = $submitedChallenge->acceptedChallenge->challenge->acceptedChallenges;
+                        $isVoted = 0;
+                        $voters = Vote::where('user_id',auth()->id())->get();
+                        if($voters->first()) {
+                            foreach($acceptedChallenges as $acceptedChallenge){
+                                foreach($voters as $voter){
+                                    if($voter->submited_challenge_id === $acceptedChallenge->submitChallenge->id){
+                                        $isVoted++;
+                                    }
                                 }
                             }
+                            if($isVoted >= 1){
+                                $data['message'] = 'You have already voted to challenger!';
+                            }
+                            $voted = $voter->where('submited_challenge_id',$submitedChallenge->id)->first();
+                            if($voted){
+                                $vote_up = $voted->vote_up = ($voted->vote_up == false) ? true : false ;
+                                ($vote_up == true) ? true : $voted->delete();
+                                $vote_down = $voted->vote_down = false ;
+                                $voted->update();
+                                $data['message'] = ($vote_up == true) ? 'Your Vote has been casted Positive on this challenge.' : 'Your Vote has been removed' ;
+                                $res = 200;
+                                $data['vote_up'] = $vote_up;
+                                $data['vote_down'] = $vote_down;
+                            }
+                        } else {
+                            $data['message'] = 'Your Vote has been casted Positive on this Challenge!';
+                            $data['vote_up'] = true;
+                            $data['vote_down'] = false;
+                            $vote = [
+                                'user_id' => auth()->id(),
+                                'submited_challenge_id' => $sub_id,
+                                'vote_up' => true,
+                            ];
+                            $this->model->create($vote);    
                         }
-                        if($isVoted >= 1){
-                            $data['message'] = 'You have already voted to challenger!';
-                        }
-                        $voted = $voter->where('submited_challenge_id',$submitedChallenge->id)->first();
-                        if($voted){
-                            $vote_up = $voted->vote_up = ($voted->vote_up == false) ? true : false ;
-                            ($vote_up == true) ? true : $voted->delete();
-                            $vote_down = $voted->vote_down = false ;
-                            $voted->update();
-                            $data['message'] = ($vote_up == true) ? 'Your Vote has been casted Positive on this challenge.' : 'Your Vote has been removed' ;
-                            $res = 200;
-                            $data['vote_up'] = $vote_up;
-                            $data['vote_down'] = $vote_down;
-                        }
-                    } else {
-                        $data['message'] = 'Your Vote has been casted Positive on this Challenge!';
-                        $data['vote_up'] = true;
-                        $data['vote_down'] = false;
-                        $vote = [
-                            'user_id' => auth()->id(),
-                            'submited_challenge_id' => $sub_id,
-                            'vote_up' => true,
-                        ];
-                        $this->model->create($vote);    
-                    }
+                    } 
                 }
             }
         }
@@ -100,42 +103,44 @@ class VoteController extends Controller
             if(auth()->id() <> $submitedChallenge->acceptedChallenge->user->id){ 
                 $data['message'] = 'The result of this Challenge is not based on Vote';
                 if($submitedChallenge->acceptedChallenge->challenge->result_type === 'vote'  ){
-                    $res = 200;
-                    $sub_id  = $submitedChallenge->id;
-                    $acceptedChallenges = $submitedChallenge->acceptedChallenge->challenge->acceptedChallenges;
-                    $isVoted = 0;
-                    $voters = Vote::where('user_id',auth()->id())->get();
-                    if($voters->first()) {
-                        foreach($acceptedChallenges as $acceptedChallenge){
-                            foreach($voters as $voter){
-                                if($voter->submited_challenge_id === $acceptedChallenge->submitChallenge->id){
-                                    $isVoted++;
+                    if(now() <= $submitedChallenge->acceptedChallenge->challenge->after_date){
+                        $res = 200;
+                        $sub_id  = $submitedChallenge->id;
+                        $acceptedChallenges = $submitedChallenge->acceptedChallenge->challenge->acceptedChallenges;
+                        $isVoted = 0;
+                        $voters = Vote::where('user_id',auth()->id())->get();
+                        if($voters->first()) {
+                            foreach($acceptedChallenges as $acceptedChallenge){
+                                foreach($voters as $voter){
+                                    if($voter->submited_challenge_id === $acceptedChallenge->submitChallenge->id){
+                                        $isVoted++;
+                                    }
                                 }
                             }
+                            if($isVoted >= 1){
+                                $data['message'] = 'You have already voted to challenger!';
+                            }
+                            $voted = $voter->where('submited_challenge_id',$submitedChallenge->id)->first();
+                            if($voted){
+                                $vote_down = $voted->vote_down = ($voted->vote_down == false) ? true : false ;
+                                ($vote_down == true) ? true : $voted->delete();
+                                $vote_up = $voted->vote_up = false ;
+                                $voted->update();
+                                $data['message'] = ($vote_down == true) ? 'Your Vote has been casted Negative on this challenge.' : 'Your Vote has been removed' ;
+                                $data['vote_up'] = $vote_up;
+                                $data['vote_down'] = $vote_down;
+                            }
+                        } else {
+                            $data['message'] = 'Your Vote has been casted Negative on this Challenge!';
+                            $data['vote_down'] = true;
+                            $data['vote_up'] = false;
+                            $vote = [
+                                'user_id' => auth()->id(),
+                                'submited_challenge_id' => $sub_id,
+                                'vote_down' => true,
+                            ];
+                            $this->model->create($vote);    
                         }
-                        if($isVoted >= 1){
-                            $data['message'] = 'You have already voted to challenger!';
-                        }
-                        $voted = $voter->where('submited_challenge_id',$submitedChallenge->id)->first();
-                        if($voted){
-                            $vote_down = $voted->vote_down = ($voted->vote_down == false) ? true : false ;
-                            ($vote_down == true) ? true : $voted->delete();
-                            $vote_up = $voted->vote_up = false ;
-                            $voted->update();
-                            $data['message'] = ($vote_down == true) ? 'Your Vote has been casted Negative on this challenge.' : 'Your Vote has been removed' ;
-                            $data['vote_up'] = $vote_up;
-                            $data['vote_down'] = $vote_down;
-                        }
-                    } else {
-                        $data['message'] = 'Your Vote has been casted Negative on this Challenge!';
-                        $data['vote_down'] = true;
-                        $data['vote_up'] = false;
-                        $vote = [
-                            'user_id' => auth()->id(),
-                            'submited_challenge_id' => $sub_id,
-                            'vote_down' => true,
-                        ];
-                        $this->model->create($vote);    
                     }
                 }
             }
